+++ src/registers.rs
@@ -106,12 +106,12 @@ pub const PACKET_CONFIG_1_CRC_WHITENING_TYPE_MASK: u8 = 0x01;
 
 // RegPacketConfig2 --------------------------------------------------------------------------------
 pub const PACKET_CONFIG_2: u8 = 0x31;
-// TODO 
+// TODO
 pub const PACKET_CONFIG_2_PAYLOAD_LENGTH: u8 = 0x07;
 
 // -------------------------------------------------------------------------------------------------
 pub const PAYLOAD_LENGTH: u8 = 0x32;
-// TODO pub const NODE_ADRS: u8 = 0x33;
+pub const NODE_ADRS: u8 = 0x33;
 // TODO pub const BROADCAST_ADRS: u8 = 0x34;
 // TODO pub const FIFO_THRESH: u8 = 0x35;
 // TODO pub const SEQ_CONFIG_1: u8 = 0x36;